% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_data.R
\docType{data}
\name{fish_data}
\alias{fish_data}
\title{fish_data dataset}
\format{
A data frame with 300 rows and 1 column
}
\usage{
fish_data
}
\description{
Data listing fish species caught
}
\author{
Naomi Tague
\describe{
\item{x}{fish species}
}
}
\keyword{datasets}
