% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_distribution.R
\name{fish_distribution}
\alias{fish_distribution}
\title{fish_distribution}
\usage{
fish_distribution(fish_data, show_plot = FALSE)
}
\arguments{
\item{fish_data}{data frame that lists fish that were caught}

\item{show_plot}{user specifies argument to show histogram.  Currently,the default is set at FALSE.}
}
\value{
A fish summary including total number of fish, rarest fish, and most common type of fish.  If show_plot is set to TRUE, a histogram of the number of fish caught by species will also be returned.
}
\description{
This function reports the total number of fish caught, the most common fish species caught, and the rarest fish species caught from an inputted .txt file listing all fish caught. User can change default settings to "TRUE" to also display a histogram of the number of fish caught per species.
}
\author{
Alex Milward, Leana Goetze, Allison Bailey
}
