% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightening.R
\name{lightening}
\alias{lightening}
\title{lightening}
\usage{
lightening(i = 0)
}
\arguments{
\item{i}{lightning activity level (ranges from 1-6)}
}
\value{
A lightening summary of lightening levels categorized.
}
\description{
This function reports the numerical lightening levels and the categorization for each lightening level.
}
\author{
Allison Bailey
}
