% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_density.R
\name{thinning}
\alias{thinning}
\title{Forest thinning function}
\usage{
thinning(day = 0, plots, original_density = 350)
}
\arguments{
\item{day}{Day forest thinning project will end. Default starting day is day 0.}

\item{plots}{Number of plots interested in thinning. Each plot is 1 ha.}

\item{original_density}{starting average tree density of all plots (trees/ha). Default density is 350 trees/ha.}
}
\value{
Number of days needed to thin all specified plots down to historic densities.
}
\description{
This function allows the user to determine how many days would be needed for a thinning project to reduce YPMC forest from current to historic conditions for a specified number of plots.
}
\examples{
thinning(plots=4)
thinning(day=0, plots=10, original_density=400)
}
\author{
Leana Goetze
}
