% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_divert.R
\name{water_diverted}
\alias{water_diverted}
\title{water_diverted}
\usage{
water_diverted(diversion)
}
\arguments{
\item{diversion}{is the amount of water diverted (cubic feet per second)}
}
\value{
Total water diverted in acre-feet
}
\description{
This function calculates total water diverted in acre-feet so long as flow is greater than 1000 cfs and water users only divert for at most 200 days per year
}
\author{
Alex Milward
}
